Refactored the cameras to their own plugin
